requirements.txt
aiofiles==22.1.0
anyio==3.6.2
argcomplete==2.0.0
asttokens==2.1.0
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
backports.cached-property==1.0.2
backports.zoneinfo==0.2.1
blis==0.7.9
boto3==1.26.2
botocore==1.29.2
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
dacite==1.6.0
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
dnspython==2.2.1
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.4
eventlet==0.33.1
exceptiongroup==1.0.4
executing==1.2.0
fiftyone==0.18.0
fiftyone-brain==0.9.2
fiftyone-db==0.4.0
filelock==3.8.0
fonttools==4.38.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
glob2==0.7
grad-cam==1.4.6
graphql-core==3.2.3
greenlet==2.0.0.post0
h11==0.12.0
h2==4.1.0
h5py==3.7.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.10.1
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
imageio==2.22.3
importlib-resources==5.10.0
inplace-abn==1.1.0
ipykernel==6.17.0
ipython==8.6.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==2.2.0
jsonref==1.0.0
jsonschema==4.16.0
jupyter_client==7.4.4
jupyter_core==4.11.2
kaleido==0.2.1
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mongoengine==0.24.2
mosestokenizer==1.2.1
motor==2.5.1
murmurhash==1.0.9
ndjson==0.3.1
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.7
numpy==1.23.4
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openfile==0.0.7
outcome==1.2.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
path==16.5.0
pathtools==0.1.2
pathy==0.10.0
patool==1.12
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
pkgutil_resolve_name==1.3.10
plotly==5.11.0
pprintpp==0.4.0
preshed==3.0.8
prettytable==3.5.0
priority==2.0.0
promise==2.3
prompt-toolkit==3.0.31
protobuf==4.21.8
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0
pydantic==1.10.2
pyecharts==1.9.1
Pygments==2.13.0
pymongo==3.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
pyvis==0.3.0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
retrying==1.3.3
revtok==0.0.3
rfc3986==1.5.0
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.1
selenium==4.6.0
sentinel==0.3.0
sentry-sdk==1.10.1
setproctitle==1.3.2
setuptools==63.4.1
shortuuid==1.0.9
simplejson==3.18.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
snapshot-selenium==0.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
sse-starlette==0.10.3
sseclient-py==1.7.2
stack-data==0.6.0
starlette==0.20.4
strawberry-graphql==0.138.1
sweeps==0.2.0
tabulate==0.9.0
tenacity==8.1.0
thinc==8.1.5
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.12
tokenizers==0.13.1
toml==0.10.2
toolwrapper==2.1.0
torch-cluster==1.6.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.15
torch-spline-conv==1.2.1
torchmetrics==0.10.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.23.1
treelib==1.6.1
trio==0.22.0
trio-websocket==0.9.2
ttach==0.0.3
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
uctools==1.3.0
universal-analytics-python3==1.1.1
urllib3==1.26.12
voxel51-eta==0.8.1
wandb==0.13.7
wasabi==0.10.1
wcwidth==0.2.5
wget==3.2
wheel==0.37.1
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
yacs==0.1.8
zipp==3.10.0
